Moved more methods to utils file
