+++ backend/src/utils/packageDataHelpers.ts
@@ -42,3 +42,35 @@ export function extractNameAndVersionFromPackageJson(packageJson: string) {
 
     return { Name, Version };
 }
+
+export function extractGitHubAttributesFromGitHubURL(urlRepo: string): { owner: string, repo: string } {
+    var owner = urlRepo.split('/')[3].trim();
+    var repo = urlRepo.split('/')[4].trim();
+
+    if (repo.includes('.git')) {
+        repo = repo.replace('.git', '');
+    }
+
+    console.log(`Owner: ${owner}`);
+    console.log(`Repo: ${repo}`);
+
+    return { owner, repo };
+}
+
+
+    // Fetch GitHub repository URL from npm API
+export async function getNpmRepoURLFromGitHubURL(url: string): Promise<string> {
+    const npmApiUrl = url.replace(/(?<=\/)www(?=\.)/, 'replicate').replace('/package', '');
+    console.log(`Fetching repository URL from npm API: ${npmApiUrl}`);
+    const npmApiResponse = await fetch(npmApiUrl);
+    const npmApiData = await npmApiResponse.json();
+
+    if (!npmApiData.repository || !npmApiData.repository.url) {
+        console.log(`Repository URL not found in npm package data for URL: ${url}`);
+        throw new ApiError('Repository URL not found in npm package data', 400);
+    }
+
+    const npmRepoUrl = npmApiData.repository.url;
+    console.log(`NPM Repository URL: ${npmRepoUrl}`);
+    return npmRepoUrl;
+} // end getNpmRepoURL
